feat(keymap)!: overhaul layout
First I updated to the Enthium V9 but after a while I came into
a conclusion that I really want a layout that works with a 36 key
keyboard. So now the base layout is a bastardised version of Enthium V9
with B, Z, and V switched around so they fit on a smaller footprint.
Also included ä on the base layer since I do write so much Finnish that
having it behind a modifier would make things cumbersome.

Also modified the symbols layer with the same idea in mind.